+++ bricktracker/app.py
@@ -17,6 +17,7 @@ from bricktracker.views.admin.database import admin_database_page
 from bricktracker.views.admin.image import admin_image_page
 from bricktracker.views.admin.instructions import admin_instructions_page
 from bricktracker.views.admin.owner import admin_owner_page
+from bricktracker.views.admin.purchase_location import admin_purchase_location_page  # noqa: E501
 from bricktracker.views.admin.retired import admin_retired_page
 from bricktracker.views.admin.status import admin_status_page
 from bricktracker.views.admin.storage import admin_storage_page
@@ -88,6 +89,7 @@ def setup_app(app: Flask) -> None:
     app.register_blueprint(admin_instructions_page)
     app.register_blueprint(admin_retired_page)
     app.register_blueprint(admin_owner_page)
+    app.register_blueprint(admin_purchase_location_page)
     app.register_blueprint(admin_status_page)
     app.register_blueprint(admin_storage_page)
     app.register_blueprint(admin_tag_page)