+++ bricktracker/record_list.py
@@ -9,6 +9,7 @@ if TYPE_CHECKING:
     from .rebrickable_set import RebrickableSet
     from .set import BrickSet
     from .set_owner import BrickSetOwner
+    from .set_purchase_location import BrickSetPurchaseLocation
     from .set_status import BrickSetStatus
     from .set_storage import BrickSetStorage
     from .set_tag import BrickSetTag
@@ -20,6 +21,7 @@ T = TypeVar(
     'BrickPart',
     'BrickSet',
     'BrickSetOwner',
+    'BrickSetPurchaseLocation',
     'BrickSetStatus',
     'BrickSetStorage',
     'BrickSetTag',