+++ bricktracker/sql/set/insert.sql
@@ -1,9 +1,11 @@
 INSERT OR IGNORE INTO "bricktracker_sets" (
     "id",
     "set",
-    "storage"
+    "storage",
+    "purchase_location"
 ) VALUES (
     :id,
     :set,
-    :storage
+    :storage,
+    :purchase_location
 )