+++ bricktracker/sql/set/metadata/purchase_location/base.sql
@@ -1,0 +1,6 @@
+SELECT
+    "bricktracker_metadata_purchase_locations"."id",
+    "bricktracker_metadata_purchase_locations"."name"
+FROM "bricktracker_metadata_purchase_locations"
+
+{% block where %}{% endblock %}