+++ bricktracker/views/admin/admin.py
@@ -10,6 +10,8 @@ from ...rebrickable_image import RebrickableImage
 from ...retired_list import BrickRetiredList
 from ...set_owner import BrickSetOwner
 from ...set_owner_list import BrickSetOwnerList
+from ...set_purchase_location import BrickSetPurchaseLocation
+from ...set_purchase_location_list import BrickSetPurchaseLocationList
 from ...set_storage import BrickSetStorage
 from ...set_storage_list import BrickSetStorageList
 from ...set_status import BrickSetStatus
@@ -36,6 +38,7 @@ def admin() -> str:
     database_version: int = -1
     instructions: BrickInstructionsList | None = None
     metadata_owners: list[BrickSetOwner] = []
+    metadata_purchase_locations: list[BrickSetPurchaseLocation] = []
     metadata_statuses: list[BrickSetStatus] = []
     metadata_storages: list[BrickSetStorage] = []
     metadata_tags: list[BrickSetTag] = []
@@ -54,6 +57,7 @@ def admin() -> str:
         instructions = BrickInstructionsList()
 
         metadata_owners = BrickSetOwnerList.list()
+        metadata_purchase_locations = BrickSetPurchaseLocationList.list()
         metadata_statuses = BrickSetStatusList.list(all=True)
         metadata_storages = BrickSetStorageList.list()
         metadata_tags = BrickSetTagList.list()
@@ -81,6 +85,7 @@ def admin() -> str:
     open_instructions = request.args.get('open_instructions', None)
     open_logout = request.args.get('open_logout', None)
     open_owner = request.args.get('open_owner', None)
+    open_purchase_location = request.args.get('open_purchase_location', None)
     open_retired = request.args.get('open_retired', None)
     open_status = request.args.get('open_status', None)
     open_storage = request.args.get('open_storage', None)
@@ -89,6 +94,7 @@ def admin() -> str:
 
     open_metadata = (
         open_owner or
+        open_purchase_location or
         open_status or
         open_storage or
         open_tag
@@ -113,6 +119,7 @@ def admin() -> str:
         database_version=database_version,
         instructions=instructions,
         metadata_owners=metadata_owners,
+        metadata_purchase_locations=metadata_purchase_locations,
         metadata_statuses=metadata_statuses,
         metadata_storages=metadata_storages,
         metadata_tags=metadata_tags,
@@ -126,12 +133,14 @@ def admin() -> str:
         open_logout=open_logout,
         open_metadata=open_metadata,
         open_owner=open_owner,
+        open_purchase_location=open_purchase_location,
         open_retired=open_retired,
         open_status=open_status,
         open_storage=open_storage,
         open_tag=open_tag,
         open_theme=open_theme,
         owner_error=request.args.get('owner_error'),
+        purchase_location_error=request.args.get('purchase_location_error'),
         retired=BrickRetiredList(),
         status_error=request.args.get('status_error'),
         storage_error=request.args.get('storage_error'),